feat: div 태그로 변경, state 추가 #122
- dialog 태그 대신 div 태그로 변경하였습니다.
- useState를 추가하여 상태값에 따라 dialog를 표시하도록 하였습니다.
